\name{colorfunctions}
\alias{colorfunctions}
\alias{reds}
\alias{blues}
\alias{greens}

\title{Color Functions}
\description{
Color functions for turning numerical variables into color codes.
}

\usage{
reds(level)
greens(level)
blues(level)
}
\arguments{
	\item{level}{Number between 0 and 1}
}
\seealso{
\code{\link{USALevelPlot}}, \code{\link[grDevices]{grey}}, \code{\link[grDevices]{rgb}}, \code{\link[grDevices]{hsv}}
}
\keyword{color}